- [x] Can emphasize go to the doc sheet mixin level? Ideally.
